requirements.txt
tensorflow
robosuite
numpy
scipy
tqdm
pyyaml
tensorboard